requirements.txt
openai
faiss-cpu
numpy
pandas
python-dotenv
tqdm
